+++ frontend/src/components/auth/components/Profile.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useAuth0 } from '@auth0/auth0-react'
-import { Drawer } from '../../common/components'
+import { Drawer } from '../../common'
 
 export const Profile: React.FC = () => {
   const { user, isAuthenticated, isLoading } = useAuth0()
