+++ frontend/src/components/common/NavigationButton.tsx
