+++ frontend/src/components/pages/planner/ProgramList.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import type Program from '../../../types/Program'
 import { Box } from '@mui/material'
-import { NavigationButton } from '../../common/components'
+import { NavigationButton } from '../../common'
 import { useAuth0 } from '@auth0/auth0-react'
 import { fetchUserPrograms } from '../../../utils/fetchUserPrograms'
 import CircularProgress from '@mui/material/CircularProgress'
