+++ frontend/src/components/pages/planner/index.ts
@@ -1,0 +1,4 @@
+import { Planner } from './Planner'
+import { ProgramList } from './ProgramList'
+
+export { Planner, ProgramList }
