+++ frontend/src/components/program/components/index.ts
@@ -1,5 +1,3 @@
 import { ProgramDetails } from './ProgramDetails'
-import { ProgramList } from './ProgramList'
-import { Planner } from './Planner'
 
-export { ProgramDetails, ProgramList, Planner }
+export { ProgramDetails }
